uv_tool_func_bot/requirements.txt
python-dotenv
agents
